Move validations up a level
